+++ backend/backend/settings/common.py
@@ -251,8 +251,24 @@ CELERY_ACCEPT_CONTENT = ["application/json"]
 CELERY_RESULT_SERIALIZER = "json"
 CELERY_TASK_SERIALIZER = "json"
 CELERY_TASK_TRACK_STARTED = True  # since 4.0
-CELERY_TASK_MAX_RETRIES = 5
-CELERY_TASK_RETRY_DELAY_SECONDS = 2
+
+# With these settings, tasks will be retried for up to a maximum of 127 minutes.
+#
+# max_wait = CELERY_TASK_RETRY_BACKOFF * sum(2 ** n for n in range(CELERY_TASK_MAX_RETRIES))
+#          = 60 * (1 + 2 + 4 + 8 + 16 + 32 + 64)
+#          = 127 minutes
+#
+# Since jitter is enabled, the actual cumulative wait can be much less than max_wait. From the doc
+# (https://docs.celeryproject.org/en/stable/userguide/tasks.html#Task.retry_jitter):
+#
+# > If this option is set to True, the delay value calculated by retry_backoff is treated as a maximum, and the actual
+# > delay value will be a random number between zero and that maximum.
+CELERY_TASK_AUTORETRY_FOR = (Exception,)  # Retry on any exception
+CELERY_TASK_MAX_RETRIES = int(os.environ.get("CELERY_TASK_MAX_RETRIES", 7))
+CELERY_TASK_RETRY_BACKOFF = int(os.environ.get("CELERY_TASK_RETRY_BACKOFF", 60))  # time in seconds
+CELERY_TASK_RETRY_BACKOFF_MAX = int(os.environ.get("CELERY_TASK_RETRY_BACKOFF_MAX", 64 * 60))
+CELERY_TASK_RETRY_JITTER = True
+
 CELERY_WORKER_CONCURRENCY = int(os.environ.get("CELERY_WORKER_CONCURRENCY", 1))
 CELERY_BROADCAST = f"{ORG_NAME}.broadcast"